refactor(api): split api.ts into typed resource clients behind a barrel (FE-010)
