--- frontend/src/lib/api.ts
@@ -1,2033 +1,0 @@
-import axios from "axios";
-import type { CaptureSourceReportPayload } from "@/lib/capture/sourceReport";
-import { getErrorMessage, getErrorStatus, isAbortError } from "@/lib/errors";
-import {
-  ActiveRecordingConflictDetail,
-  CalendarConnection,
-  CalendarDashboardSummary,
-  CalendarOverview,
-  CalendarProvider,
-  CalendarProviderConfigUpdate,
-  CalendarProviderStatus,
-  Recording,
-  RecordingCaptureLifecycleResponse,
-  RecordingId,
-  RecordingInitResponse,
-  RecordingsCalendar,
-  GlobalSpeaker,
-  Settings,
-  Tag,
-  TranscriptSegment,
-  VoiceprintExtractResult,
-  VoiceprintApplyResult,
-  BatchVoiceprintResponse,
-  RecordingSpeaker,
-  ChatMessage,
-  User,
-  Invitation,
-  AdminHealthStatus,
-  SystemModelStatus,
-  VersionInfo,
-  PeopleTag,
-  SpeakerSegment,
-  SegmentSelection,
-  TranscriptSpeakerAssignment,
-  TranscriptUtteranceList,
-  UserTask,
-  ReprocessRequest,
-  CalendarEventLink,
-  RecordingInfo,
-  AsyncTaskStatus,
-  DownloadProgress,
-  LanguageRegistry,
-} from "@/types";
-
-export const API_BASE_URL = process.env.NEXT_PUBLIC_API_URL
-  ? `${process.env.NEXT_PUBLIC_API_URL}/v1`
-  : "https://localhost:14443/api/v1";
-const FIRST_RUN_PASSWORD_AUTH_SCHEME = "Bootstrap";
-const FORCE_PASSWORD_CHANGE_REDIRECT = "/settings?tab=account&forcePasswordChange=1";
-
-type ValidationErrorItem = {
-  loc?: unknown;
-  msg?: unknown;
-};
-
-const formatValidationDetail = (detail: unknown): string | unknown => {
-  if (!Array.isArray(detail)) {
-    return detail;
-  }
-
-  const messages = detail
-    .map((item) => {
-      if (!item || typeof item !== "object") {
-        return null;
-      }
-
-      const { loc, msg } = item as ValidationErrorItem;
-      if (typeof msg !== "string") {
-        return null;
-      }
-
-      if (!Array.isArray(loc) || typeof loc[loc.length - 1] !== "string") {
-        return msg;
-      }
-
-      const fieldName = String(loc[loc.length - 1]).replace(/_/g, " ");
-      return `${fieldName}: ${msg}`;
-    })
-    .filter((message): message is string => Boolean(message));
-
-  if (messages.length === 0) {
-    return detail;
-  }
-
-  return messages.join(" ");
-};
-
-const buildFirstRunRequestConfig = (bootstrapPassword?: string) => {
-  if (!bootstrapPassword) {
-    return {};
-  }
-
-  return {
-    headers: {
-      Authorization: `${FIRST_RUN_PASSWORD_AUTH_SCHEME} ${bootstrapPassword}`,
-    },
-  };
-};
-
-const api = axios.create({
-  baseURL: API_BASE_URL,
-  maxRedirects: 0,
-  withCredentials: true,
-  headers: {
-    "Content-Type": "application/json",
-  },
-});
-
-api.interceptors.request.use((config) => {
-  return config;
-});
-
-api.interceptors.response.use(
-  (response) => response,
-  (error) => {
-    if (
-      error.response?.data &&
-      typeof error.response.data === "object" &&
-      !Array.isArray(error.response.data)
-    ) {
-      error.response.data.detail = formatValidationDetail(
-        error.response.data.detail,
-      );
-    }
-
-    if (error.response && error.response.status === 401) {
-      if (
-        typeof window !== "undefined" &&
-        !window.location.pathname.includes("/login") &&
-        !window.location.pathname.includes("/setup") &&
-        !window.location.pathname.includes("/register")
-      ) {
-        window.location.href = "/login";
-      }
-    }
-
-    if (
-      error.response &&
-      error.response.status === 403 &&
-      error.response.data?.detail === "Password change required"
-    ) {
-      if (
-        typeof window !== "undefined" &&
-        !window.location.pathname.startsWith("/settings")
-      ) {
-        window.location.href = FORCE_PASSWORD_CHANGE_REDIRECT;
-      }
-    }
-
-    return Promise.reject(error);
-  },
-);
-
-
-export const login = async (
-  username: string,
-  password: string,
-): Promise<{
-  force_password_change: boolean;
-  is_superuser: boolean;
-  username: string;
-}> => {
-  const formData = new FormData();
-  formData.append("username", username);
-  formData.append("password", password);
-
-  const response = await api.post<{
-    force_password_change: boolean;
-    is_superuser: boolean;
-    username: string;
-  }>("/login/session", formData, {
-    headers: {
-      "Content-Type": "multipart/form-data",
-    },
-  });
-  return response.data;
-};
-
-export const logout = async (): Promise<void> => {
-  try {
-    await api.post("/login/logout");
-
-    } catch (error: unknown) {
-    console.error("Logout failed:", error);
-  } finally {
-    if (typeof window !== "undefined") {
-      window.location.href = "/login";
-    }
-  }
-};
-
-export const getCurrentUser = async (): Promise<User> => {
-  const response = await api.get<User>("/users/me");
-  return response.data;
-};
-
-export interface RecordingFilters {
-  q?: string;
-  start_date?: string;
-  end_date?: string;
-  speaker_ids?: number[];
-  tag_ids?: number[];
-  include_archived?: boolean;
-  include_deleted?: boolean;
-  only_archived?: boolean;
-  only_deleted?: boolean;
-}
-
-export const getRecordings = async (
-  filters?: RecordingFilters,
-): Promise<Recording[]> => {
-  const params = new URLSearchParams();
-
-  if (filters) {
-    if (filters.q) params.append("q", filters.q);
-    if (filters.start_date) params.append("start_date", filters.start_date);
-    if (filters.end_date) params.append("end_date", filters.end_date);
-    if (filters.speaker_ids) {
-      filters.speaker_ids.forEach((id) =>
-        params.append("speaker_ids", id.toString()),
-      );
-    }
-    if (filters.tag_ids) {
-      filters.tag_ids.forEach((id) => params.append("tag_ids", id.toString()));
-    }
-    if (filters.include_archived) params.append("include_archived", "true");
-    if (filters.include_deleted) params.append("include_deleted", "true");
-    if (filters.only_archived) params.append("only_archived", "true");
-    if (filters.only_deleted) params.append("only_deleted", "true");
-  }
-
-  const response = await api.get<Recording[]>(
-    `/recordings/?${params.toString()}`,
-  );
-  return response.data;
-};
-
-export const getRecording = async (id: RecordingId): Promise<Recording> => {
-  const response = await api.get<Recording>(`/recordings/${id}`);
-  return response.data;
-};
-
-export const initRecording = async (name?: string): Promise<RecordingInitResponse> => {
-  const params = new URLSearchParams();
-
-  if (name?.trim()) {
-    params.set("name", name.trim());
-  }
-
-  const suffix = params.toString();
-  const response = await api.post<RecordingInitResponse>(
-    `/recordings/init${suffix ? `?${suffix}` : ""}`,
-  );
-  return response.data;
-};
-
-export const pauseRecordingCapture = async (
-  recordingId: RecordingId,
-): Promise<RecordingCaptureLifecycleResponse> => {
-  const response = await api.post<RecordingCaptureLifecycleResponse>(
-    `/recordings/${recordingId}/pause`,
-  );
-  return response.data;
-};
-
-export const reportRecordingCaptureSources = async (
-  recordingId: RecordingId,
-  payload: CaptureSourceReportPayload,
-): Promise<void> => {
-  await api.post(`/recordings/${recordingId}/capture-source-report`, payload);
-};
-
-export const resumeRecordingCapture = async (
-  recordingId: RecordingId,
-): Promise<RecordingCaptureLifecycleResponse> => {
-  const response = await api.post<RecordingCaptureLifecycleResponse>(
-    `/recordings/${recordingId}/resume`,
-  );
-  return response.data;
-};
-
-export const finalizeRecordingCapture = async (
-  recordingId: RecordingId,
-): Promise<Recording> => {
-  const response = await api.post<Recording>(`/recordings/${recordingId}/finalize`);
-  return response.data;
-};
-
-export const discardRecordingCapture = async (
-  recordingId: RecordingId,
-  discardReason?: string,
-): Promise<void> => {
-  const params = new URLSearchParams();
-  if (discardReason?.trim()) {
-    params.set("reason", discardReason.trim());
-  }
-
-  const suffix = params.toString();
-  await api.post(`/recordings/${recordingId}/discard${suffix ? `?${suffix}` : ""}`);
-};
-
-const resolveRecordingSegmentExtension = (contentType: string | undefined) => {
-  const normalized = (contentType || "").split(";", 1)[0].trim().toLowerCase();
-  switch (normalized) {
-    case "audio/wav":
-      return "wav";
-    case "audio/ogg":
-      return "ogg";
-    case "audio/mp4":
-      return "m4a";
-    case "audio/webm":
-    default:
-      return "webm";
-  }
-};
-
-export const uploadRecordingSegment = async (
-  recordingId: RecordingId,
-  sequence: number,
-  blob: Blob,
-  filename = `${sequence}.${resolveRecordingSegmentExtension(blob.type)}`,
-): Promise<{ status: string; segment: number }> => {
-  const formData = new FormData();
-  formData.append("file", blob, filename);
-
-  const response = await api.post<{ status: string; segment: number }>(
-    `/recordings/${recordingId}/segment?sequence=${sequence}`,
-    formData,
-    {
-      headers: {
-        "Content-Type": "multipart/form-data",
-      },
-    },
-  );
-  return response.data;
-};
-
-export const getPausedRecordings = async (): Promise<Recording[]> => {
-  const response = await api.get<Recording[]>(
-    "/recordings/?status=PAUSED&user=me",
-  );
-  return response.data;
-};
-
-export const isActiveRecordingConflictDetail = (
-  detail: unknown,
-): detail is ActiveRecordingConflictDetail => {
-  if (!detail || typeof detail !== "object") {
-    return false;
-  }
-
-  const candidate = detail as Partial<ActiveRecordingConflictDetail>;
-  return (
-    candidate.code === "active_recording_exists" &&
-    typeof candidate.message === "string" &&
-    typeof candidate.recording_id === "string" &&
-    typeof candidate.status === "string"
-  );
-};
-
-export const linkRecordingCalendarEvent = async (
-  recordingId: RecordingId,
-  calendarEventId: number | null,
-): Promise<Recording> => {
-  const response = await api.put<Recording>(
-    `/recordings/${recordingId}/calendar-event`,
-    { calendar_event_id: calendarEventId },
-  );
-  return response.data;
-};
-
-export const getRecordingCalendarEventCandidates = async (
-  recordingId: RecordingId,
-): Promise<CalendarEventLink[]> => {
-  const response = await api.get<CalendarEventLink[]>(
-    `/recordings/${recordingId}/calendar-event/candidates`,
-  );
-  return response.data;
-};
-
-export type UserTaskStatusFilter =
-  | "active"
-  | "open"
-  | "completed"
-  | "archived"
-  | "all";
-
-export const getUserTasks = async (
-  status: UserTaskStatusFilter = "active",
-): Promise<UserTask[]> => {
-  const response = await api.get<UserTask[]>("/tasks/", {
-    params: { status },
-  });
-  return response.data;
-};
-
-export const createUserTask = async (data: {
-  title: string;
-  body?: string | null;
-  due_at?: string | null;
-  tag_ids?: number[];
-  recording_ids?: string[];
-}): Promise<UserTask> => {
-  const response = await api.post<UserTask>("/tasks/", data);
-  return response.data;
-};
-
-export const updateUserTask = async (
-  taskId: number,
-  data: {
-    title?: string;
-    body?: string | null;
-    due_at?: string | null;
-    completed?: boolean;
-    archived?: boolean;
-    tag_ids?: number[];
-    recording_ids?: string[];
-  },
-): Promise<UserTask> => {
-  const response = await api.patch<UserTask>(`/tasks/${taskId}`, data);
-  return response.data;
-};
-
-export const deleteUserTask = async (taskId: number): Promise<void> => {
-  await api.delete(`/tasks/${taskId}`);
-};
-
-export const deleteRecording = async (id: RecordingId): Promise<void> => {
-  await api.delete(`/recordings/${id}`);
-};
-
-export const renameRecording = async (
-  id: RecordingId,
-  name: string,
-): Promise<Recording> => {
-  const response = await api.patch<Recording>(`/recordings/${id}`, { name });
-  return response.data;
-};
-
-export const reprocessRecording = async (
-  id: RecordingId,
-  body: ReprocessRequest,
-): Promise<Recording> => {
-  const response = await api.post<Recording>(
-    `/recordings/${id}/reprocess`,
-    body,
-  );
-  return response.data;
-};
-
-export const cancelProcessing = async (id: RecordingId): Promise<Recording> => {
-  const response = await api.post<Recording>(`/recordings/${id}/cancel`);
-  return response.data;
-};
-
-export const getRecordingStreamUrl = (id: RecordingId): string => {
-  return `${API_BASE_URL}/recordings/${id}/stream`;
-};
-
-// Speaker Management
-export interface GlobalSpeakerFilters {
-  q?: string;
-  tags?: number[];
-}
-
-export const getGlobalSpeakers = async (
-  filters?: GlobalSpeakerFilters,
-): Promise<GlobalSpeaker[]> => {
-  const params = new URLSearchParams();
-  if (filters) {
-    if (filters.q) params.append("q", filters.q);
-    if (filters.tags) {
-      filters.tags.forEach((id) => params.append("tags", id.toString()));
-    }
-  }
-  const response = await api.get<GlobalSpeaker[]>(
-    `/speakers/?${params.toString()}`,
-  );
-  return response.data;
-};
-
-export const createGlobalSpeaker = async (
-  data: Partial<GlobalSpeaker> & { tag_ids?: number[] },
-): Promise<GlobalSpeaker> => {
-  const response = await api.post<GlobalSpeaker>("/speakers/", data);
-  return response.data;
-};
-
-export const updateGlobalSpeaker = async (
-  id: number,
-  data: Partial<GlobalSpeaker> & { tag_ids?: number[] },
-): Promise<GlobalSpeaker> => {
-  const response = await api.put<GlobalSpeaker>(`/speakers/${id}`, data);
-  return response.data;
-};
-
-export const mergeSpeakers = async (
-  sourceId: number,
-  targetId: number,
-): Promise<GlobalSpeaker> => {
-  const response = await api.post<GlobalSpeaker>("/speakers/merge", {
-    source_speaker_id: sourceId,
-    target_speaker_id: targetId,
-  });
-  return response.data;
-};
-
-export const deleteGlobalSpeakerEmbedding = async (
-  id: number,
-): Promise<void> => {
-  await api.delete(`/speakers/${id}/embedding`);
-};
-
-export const deleteGlobalSpeaker = async (id: number): Promise<void> => {
-  await api.delete(`/speakers/${id}`);
-};
-
-export const updateSpeaker = async (
-  recordingId: RecordingId,
-  diarizationLabel: string,
-  newName: string,
-): Promise<void> => {
-  await api.put(`/speakers/recordings/${recordingId}`, {
-    diarization_label: diarizationLabel,
-    global_speaker_name: newName,
-  });
-};
-
-export const updateSpeakerColor = async (
-  recordingId: RecordingId,
-  label: string,
-  color: string,
-): Promise<void> => {
-  await api.put(
-    `/speakers/recordings/${recordingId}/speakers/${encodeURIComponent(label)}/color`,
-    {
-      color,
-    },
-  );
-};
-
-export const acceptSpeakerNameSuggestion = async (
-  recordingId: RecordingId,
-  diarizationLabel: string,
-): Promise<void> => {
-  await api.post(
-    `/speakers/recordings/${recordingId}/speakers/${encodeURIComponent(diarizationLabel)}/suggestions/accept`,
-  );
-};
-
-export const rejectSpeakerNameSuggestion = async (
-  recordingId: RecordingId,
-  diarizationLabel: string,
-): Promise<void> => {
-  await api.post(
-    `/speakers/recordings/${recordingId}/speakers/${encodeURIComponent(diarizationLabel)}/suggestions/reject`,
-  );
-};
-
-export const updateTranscriptSegmentSpeaker = async (
-  recordingId: RecordingId,
-  segmentIndex: number,
-  assignment: TranscriptSpeakerAssignment,
-): Promise<void> => {
-  await api.put(`/transcripts/${recordingId}/segments/${segmentIndex}`, {
-    new_speaker_name: assignment.name,
-    global_speaker_id: assignment.globalSpeakerId,
-    diarization_label: assignment.diarizationLabel,
-  });
-};
-
-export const updateTranscriptUtteranceSpeaker = async (
-  recordingId: RecordingId,
-  utteranceId: string,
-  assignment: TranscriptSpeakerAssignment,
-): Promise<void> => {
-  await api.patch(`/transcripts/${recordingId}/utterances/${utteranceId}/speaker`, {
-    new_speaker_name: assignment.name,
-    global_speaker_id: assignment.globalSpeakerId,
-    diarization_label: assignment.diarizationLabel,
-    scope: assignment.scope,
-  });
-};
-
-export const getTranscriptUtterances = async (
-  recordingId: RecordingId,
-  afterRevision?: number,
-): Promise<TranscriptUtteranceList> => {
-  const params = new URLSearchParams();
-
-  if (afterRevision !== undefined) {
-    params.set("after_revision", String(afterRevision));
-  }
-
-  const suffix = params.toString();
-  const response = await api.get<TranscriptUtteranceList>(
-    `/transcripts/${recordingId}/utterances${suffix ? `?${suffix}` : ""}`,
-  );
-  return response.data;
-};
-
-// Tags
-export const getTags = async (): Promise<Tag[]> => {
-  const response = await api.get<Tag[]>("/tags/");
-  return response.data;
-};
-
-export const createTag = async (
-  name: string,
-  color?: string,
-  parent_id?: number,
-): Promise<Tag> => {
-  const response = await api.post<Tag>("/tags/", { name, color, parent_id });
-  return response.data;
-};
-
-export const updateTag = async (
-  tagId: number,
-  data: { name?: string; color?: string; parent_id?: number | null },
-): Promise<Tag> => {
-  const response = await api.patch<Tag>(`/tags/${tagId}`, data);
-  return response.data;
-};
-
-export const addTagToRecording = async (
-  recordingId: RecordingId,
-  tagName: string,
-): Promise<void> => {
-  await api.post(`/tags/recordings/${recordingId}`, { name: tagName });
-};
-
-export const removeTagFromRecording = async (
-  recordingId: RecordingId,
-  tagName: string,
-): Promise<void> => {
-  await api.delete(`/tags/recordings/${recordingId}/${tagName}`);
-};
-
-export const deleteTag = async (id: number): Promise<void> => {
-  await api.delete(`/tags/${id}`);
-};
-
-// People Tags
-export const getPeopleTags = async (): Promise<PeopleTag[]> => {
-  const response = await api.get<PeopleTag[]>("/people-tags/");
-  return response.data;
-};
-
-export const createPeopleTag = async (
-  name: string,
-  color?: string,
-  parent_id?: number,
-): Promise<PeopleTag> => {
-  const response = await api.post<PeopleTag>("/people-tags/", {
-    name,
-    color,
-    parent_id,
-  });
-  return response.data;
-};
-
-export const updatePeopleTag = async (
-  id: number,
-  data: { name?: string; color?: string; parent_id?: number },
-): Promise<PeopleTag> => {
-  const response = await api.patch<PeopleTag>(`/people-tags/${id}`, data);
-  return response.data;
-};
-
-export const deletePeopleTag = async (id: number): Promise<void> => {
-  await api.delete(`/people-tags/${id}`);
-};
-
-// Archive & Delete Management
-export const archiveRecording = async (id: RecordingId): Promise<Recording> => {
-  const response = await api.post<Recording>(`/recordings/${id}/archive`);
-  return response.data;
-};
-
-export const restoreRecording = async (id: RecordingId): Promise<Recording> => {
-  const response = await api.post<Recording>(`/recordings/${id}/restore`);
-  return response.data;
-};
-
-export const softDeleteRecording = async (id: RecordingId): Promise<Recording> => {
-  const response = await api.post<Recording>(`/recordings/${id}/soft-delete`);
-  return response.data;
-};
-
-export const permanentlyDeleteRecording = async (id: RecordingId): Promise<void> => {
-  await api.delete(`/recordings/${id}/permanent`);
-};
-
-// Settings
-export const getSettings = async (): Promise<Settings> => {
-  const response = await api.get<Settings>("/settings");
-  return response.data;
-};
-
-export const updateSettings = async (settings: Settings): Promise<Settings> => {
-  const response = await api.post<Settings>("/settings", settings);
-  return response.data;
-};
-
-export const getLanguageOptions = async (): Promise<LanguageRegistry> => {
-  const response = await api.get<LanguageRegistry>("/settings/languages");
-  return response.data;
-};
-
-export const getCalendarOverview = async (): Promise<CalendarOverview> => {
-  const response = await api.get<CalendarOverview>("/calendar");
-  return response.data;
-};
-
-export const getCalendarDashboardSummary = async (
-  month: string,
-  timeZone?: string,
-): Promise<CalendarDashboardSummary> => {
-  const params = new URLSearchParams();
-  params.set("month", month);
-  if (timeZone) {
-    params.set("timezone", timeZone);
-  }
-
-  const response = await api.get<CalendarDashboardSummary>(
-    `/calendar/dashboard?${params.toString()}`,
-  );
-  return response.data;
-};
-
-export const getRecordingsCalendar = async (
-  month: string,
-  timeZone?: string,
-): Promise<RecordingsCalendar> => {
-  const params = new URLSearchParams();
-  params.set("month", month);
-  if (timeZone) {
-    params.set("timezone", timeZone);
-  }
-
-  const response = await api.get<RecordingsCalendar>(
-    `/recordings/calendar?${params.toString()}`,
-  );
-  return response.data;
-};
-
-export const startCalendarAuthorisation = async (
-  provider: CalendarProvider,
-): Promise<{ authorisation_url: string }> => {
-  const response = await api.post<{ authorisation_url: string }>(
-    `/calendar/oauth/${provider}/start`,
-  );
-  return response.data;
-};
-
-export const getCalendarAuthorisationStartUrl = (
-  provider: CalendarProvider,
-): string => {
-  const path = `${API_BASE_URL}/calendar/oauth/${provider}/start`;
-  if (typeof window === "undefined") {
-    return path;
-  }
-  return new URL(path, window.location.origin).toString();
-};
-
-export const updateCalendarSelection = async (
-  connectionId: number,
-  selectedCalendarIds: number[],
-): Promise<CalendarConnection> => {
-  const response = await api.put<CalendarConnection>(
-    `/calendar/connections/${connectionId}/calendars`,
-    { selected_calendar_ids: selectedCalendarIds },
-  );
-  return response.data;
-};
-
-export const updateCalendarColor = async (
-  connectionId: number,
-  calendarId: number,
-  color: string | null,
-): Promise<CalendarConnection> => {
-  const response = await api.put<CalendarConnection>(
-    `/calendar/connections/${connectionId}/calendars/${calendarId}/color`,
-    { colour: color },
-  );
-  return response.data;
-};
-
-export const syncCalendarConnection = async (
-  connectionId: number,
-): Promise<CalendarConnection> => {
-  const response = await api.post<CalendarConnection>(
-    `/calendar/connections/${connectionId}/sync`,
-  );
-  return response.data;
-};
-
-export const disconnectCalendarConnection = async (
-  connectionId: number,
-): Promise<void> => {
-  await api.delete(`/calendar/connections/${connectionId}`);
-};
-
-export const getCalendarProviderStatuses = async (): Promise<
-  CalendarProviderStatus[]
-> => {
-  const response = await api.get<CalendarProviderStatus[]>(
-    "/calendar/admin/providers",
-  );
-  return response.data;
-};
-
-export const updateCalendarProviderConfiguration = async (
-  provider: CalendarProvider,
-  payload: CalendarProviderConfigUpdate,
-): Promise<CalendarProviderStatus> => {
-  const response = await api.put<CalendarProviderStatus>(
-    `/calendar/admin/providers/${provider}`,
-    payload,
-  );
-  return response.data;
-};
-
-// Personal Dictionary
-export const getPersonalDictionary = async (): Promise<string[]> => {
-  const response = await api.get<string[]>("/settings/personal-dictionary");
-  return response.data;
-};
-
-export const addPersonalDictionaryWord = async (word: string): Promise<void> => {
-  await api.post("/settings/personal-dictionary", { word });
-};
-
-export const removePersonalDictionaryWord = async (word: string): Promise<void> => {
-  await api.delete(`/settings/personal-dictionary/${encodeURIComponent(word)}`);
-};
-
-// Spellcheck Ignored Words
-export const getSpellcheckIgnoredWords = async (): Promise<string[]> => {
-  const response = await api.get<string[]>("/settings/spellcheck-ignored");
-  return response.data;
-};
-
-export const addSpellcheckIgnoredWord = async (word: string): Promise<void> => {
-  await api.post("/settings/spellcheck-ignored", { word });
-};
-
-export const removeSpellcheckIgnoredWord = async (word: string): Promise<void> => {
-  await api.delete(`/settings/spellcheck-ignored/${encodeURIComponent(word)}`);
-};
-
-// Transcript Text
-export const updateTranscriptSegmentText = async (
-  recordingId: RecordingId,
-  segmentIndex: number,
-  text: string,
-): Promise<void> => {
-  await api.put(`/transcripts/${recordingId}/segments/${segmentIndex}/text`, {
-    text,
-  });
-};
-
-export const updateTranscriptUtteranceText = async (
-  recordingId: RecordingId,
-  utteranceId: string,
-  text: string,
-  expectedRevision?: number,
-): Promise<void> => {
-  await api.patch(`/transcripts/${recordingId}/utterances/${utteranceId}/text`, {
-    text,
-    expected_revision: expectedRevision,
-  });
-};
-
-export const findAndReplace = async (
-  recordingId: RecordingId,
-  find: string,
-  replace: string,
-  options: { caseSensitive?: boolean; useRegex?: boolean } = {},
-): Promise<void> => {
-  await api.post(`/transcripts/${recordingId}/replace`, {
-    find_text: find,
-    replace_text: replace,
-    case_sensitive: options.caseSensitive ?? false,
-    use_regex: options.useRegex ?? false,
-  });
-};
-
-export type ExportContentType = "transcript" | "notes" | "both" | "audio";
-export type ExportFormat = "txt" | "pdf" | "docx";
-
-export const exportAudio = async (
-  recordingId: RecordingId,
-  recordingName: string,
-): Promise<void> => {
-  try {
-    const response = await api.get(`/recordings/${recordingId}/stream`, {
-      responseType: "blob",
-    });
-
-    const url = window.URL.createObjectURL(new Blob([response.data], { type: "audio/mpeg" }));
-    const link = document.createElement("a");
-    link.href = url;
-
-    // Sanitize filename similar to backend
-    const sanitizedName = recordingName.replace(/[^a-zA-Z0-9 \-_.]/g, "").trim();
-    const filename = `${sanitizedName || `recording_${recordingId}`}.mp3`;
-
-    link.setAttribute("download", filename);
-    document.body.appendChild(link);
-    link.click();
-    link.remove();
-    window.URL.revokeObjectURL(url);
-
-    } catch (error: unknown) {
-    console.error("[exportAudio] Error during audio export:", error);
-    throw error;
-  }
-};
-
-export const exportContent = async (
-  recordingId: RecordingId,
-  contentType: ExportContentType = "transcript",
-  format: ExportFormat = "txt",
-): Promise<void> => {
-  try {
-    const response = await api.get(`/transcripts/${recordingId}/export`, {
-      params: {
-        content_type: contentType,
-        export_format: format,
-      },
-      responseType: "blob",
-    });
-
-    // Create a link and click it to download
-    const url = window.URL.createObjectURL(new Blob([response.data]));
-    const link = document.createElement("a");
-    link.href = url;
-
-    // Extract filename from header if possible, or generate one
-    const contentDisposition = response.headers["content-disposition"];
-    let filename = `export-${recordingId}.${format}`;
-
-    if (contentDisposition) {
-      // Try to match filename="name"
-      const filenameMatch = contentDisposition.match(/filename="([^"]+)"/);
-      if (filenameMatch && filenameMatch.length === 2) {
-        filename = filenameMatch[1];
-      } else {
-        // Try to match filename=name
-        const filenameSimpleMatch =
-          contentDisposition.match(/filename=([^;]+)/);
-        if (filenameSimpleMatch && filenameSimpleMatch.length === 2) {
-          filename = filenameSimpleMatch[1].trim();
-        }
-      }
-    }
-
-    link.setAttribute("download", filename);
-    document.body.appendChild(link);
-    link.click();
-    link.remove();
-    window.URL.revokeObjectURL(url);
-
-    } catch (error: unknown) {
-    console.error("[exportContent] Error during export:", error);
-    throw error;
-  }
-};
-
-// Meeting Notes API
-export const getNotes = async (
-  recordingId: RecordingId,
-): Promise<{ notes: string | null }> => {
-  const response = await api.get<{ notes: string | null }>(
-    `/transcripts/${recordingId}/notes`,
-  );
-  return response.data;
-};
-
-export const updateNotes = async (
-  recordingId: RecordingId,
-  notes: string,
-): Promise<{ notes: string; status: string }> => {
-  const response = await api.put<{ notes: string; status: string }>(
-    `/transcripts/${recordingId}/notes`,
-    { notes },
-  );
-  return response.data;
-};
-
-export const getUserNotes = async (
-  recordingId: RecordingId,
-): Promise<{ user_notes: string | null }> => {
-  const response = await api.get<{ user_notes: string | null }>(
-    `/transcripts/${recordingId}/user-notes`,
-  );
-  return response.data;
-};
-
-export const updateUserNotes = async (
-  recordingId: RecordingId,
-  userNotes: string,
-): Promise<{ user_notes: string | null; status: string }> => {
-  const response = await api.put<{ user_notes: string | null; status: string }>(
-    `/transcripts/${recordingId}/user-notes`,
-    { user_notes: userNotes },
-  );
-  return response.data;
-};
-
-export const updateMeetingEdgeFocus = async (
-  recordingId: RecordingId,
-  meetingEdgeFocus: string,
-): Promise<{ meeting_edge_focus: string | null; status: string }> => {
-  const response = await api.put<{
-    meeting_edge_focus: string | null;
-    status: string;
-  }>(`/transcripts/${recordingId}/meeting-edge-focus`, {
-    meeting_edge_focus: meetingEdgeFocus,
-  });
-  return response.data;
-};
-
-export const generateNotes = async (
-  recordingId: RecordingId,
-): Promise<{
-  status: string;
-  notes_status?: string;
-  error_message?: string | null;
-  message?: string;
-}> => {
-  const response = await api.post<{
-    status: string;
-    notes_status?: string;
-    error_message?: string | null;
-    message?: string;
-  }>(
-    `/transcripts/${recordingId}/notes/generate`,
-  );
-  return response.data;
-};
-
-export const findAndReplaceNotes = async (
-  recordingId: RecordingId,
-  find: string,
-  replace: string,
-  options: { caseSensitive?: boolean; useRegex?: boolean } = {},
-): Promise<void> => {
-  // Use the main replace endpoint since it applies to both transcript and notes
-  await api.post(`/transcripts/${recordingId}/replace`, {
-    find_text: find,
-    replace_text: replace,
-    case_sensitive: options.caseSensitive ?? false,
-    use_regex: options.useRegex ?? false,
-  });
-};
-
-export const mergeRecordingSpeakers = async (
-  recordingId: RecordingId,
-  targetSpeakerLabel: string,
-  sourceSpeakerLabel: string,
-): Promise<Recording> => {
-  const response = await api.post<Recording>(
-    `/speakers/recordings/${recordingId}/merge`,
-    {
-      target_speaker_label: targetSpeakerLabel,
-      source_speaker_label: sourceSpeakerLabel,
-    },
-  );
-  return response.data;
-};
-
-export const deleteRecordingSpeaker = async (
-  recordingId: RecordingId,
-  diarizationLabel: string,
-): Promise<void> => {
-  await api.delete(
-    `/speakers/recordings/${recordingId}/speakers/${encodeURIComponent(diarizationLabel)}`,
-  );
-};
-
-export const promoteToGlobalSpeaker = async (
-  recordingId: RecordingId,
-  diarizationLabel: string,
-): Promise<RecordingSpeaker> => {
-  const response = await api.post<RecordingSpeaker>(
-    `/speakers/recordings/${recordingId}/speakers/${encodeURIComponent(diarizationLabel)}/promote`,
-  );
-  return response.data;
-};
-
-export const updateTranscriptSegments = async (
-  recordingId: RecordingId,
-  segments: TranscriptSegment[],
-): Promise<void> => {
-  await api.put(`/transcripts/${recordingId}/segments`, { segments });
-};
-
-export interface ImportAudioOptions {
-  name?: string;
-  recordedAt?: Date;
-  onUploadProgress?: (progress: number) => void;
-}
-
-export const importAudio = async (
-  file: File,
-  options?: ImportAudioOptions,
-): Promise<Recording> => {
-  // Use chunked upload for all files to ensure reliability and bypass Cloudflare limits
-  // Chunk size: 10MB
-  const CHUNK_SIZE = 10 * 1024 * 1024;
-  const totalChunks = Math.ceil(file.size / CHUNK_SIZE);
-
-  // 1. Initialize Import
-  const initParams = new URLSearchParams();
-  initParams.append("filename", file.name);
-  if (options?.name) initParams.append("name", options.name);
-  if (options?.recordedAt)
-    initParams.append("recorded_at", options.recordedAt.toISOString());
-
-  const initResponse = await api.post<Recording>(
-    `/recordings/import/chunked/init?${initParams.toString()}`,
-  );
-  const recording = initResponse.data;
-
-  // 2. Upload Chunks
-  for (let i = 0; i < totalChunks; i++) {
-    const start = i * CHUNK_SIZE;
-    const end = Math.min(start + CHUNK_SIZE, file.size);
-    const chunk = file.slice(start, end);
-
-    const formData = new FormData();
-    formData.append("file", chunk);
-
-    // Chunked imports keep their existing import-part sequence; browser live capture uses 0-based segments.
-    await api.post(
-      `/recordings/import/chunked/segment?recording_id=${recording.id}&sequence=${i + 1}`,
-      formData,
-      {
-        headers: { "Content-Type": "multipart/form-data" },
-      },
-    );
-
-    // Update progress
-    if (options?.onUploadProgress) {
-      const progress = Math.round(((i + 1) / totalChunks) * 100);
-      // Cap at 99% until finalized
-      options.onUploadProgress(Math.min(progress, 99));
-    }
-  }
-
-  // 3. Finalize Import
-  const finalizeResponse = await api.post<Recording>(
-    `/recordings/import/chunked/finalize?recording_id=${recording.id}`,
-  );
-
-  if (options?.onUploadProgress) {
-    options.onUploadProgress(100);
-  }
-
-  return finalizeResponse.data;
-};
-
-export const getSupportedAudioFormats = (): string[] => {
-  return [
-    ".wav",
-    ".mp3",
-    ".m4a",
-    ".aac",
-    ".webm",
-    ".ogg",
-    ".flac",
-    ".mp4",
-    ".wma",
-    ".opus",
-  ];
-};
-
-
-
-// Voiceprint Management
-export const extractVoiceprint = async (
-  recordingId: RecordingId,
-  diarizationLabel: string,
-): Promise<VoiceprintExtractResult> => {
-  const response = await api.post<VoiceprintExtractResult>(
-    `/speakers/recordings/${recordingId}/speakers/${encodeURIComponent(diarizationLabel)}/voiceprint/extract`,
-  );
-  return response.data;
-};
-
-export type VoiceprintAction =
-  | "create_new"
-  | "link_existing"
-  | "local_only"
-  | "force_link";
-
-export const applyVoiceprintAction = async (
-  recordingId: RecordingId,
-  diarizationLabel: string,
-  action: VoiceprintAction,
-  options?: { globalSpeakerId?: number; newSpeakerName?: string },
-): Promise<VoiceprintApplyResult> => {
-  const response = await api.post<VoiceprintApplyResult>(
-    `/speakers/recordings/${recordingId}/speakers/${encodeURIComponent(diarizationLabel)}/voiceprint/apply`,
-    {
-      action,
-      global_speaker_id: options?.globalSpeakerId,
-      new_speaker_name: options?.newSpeakerName,
-    },
-  );
-  return response.data;
-};
-
-export const deleteVoiceprint = async (
-  recordingId: RecordingId,
-  diarizationLabel: string,
-): Promise<void> => {
-  await api.delete(
-    `/speakers/recordings/${recordingId}/speakers/${encodeURIComponent(diarizationLabel)}/voiceprint`,
-  );
-};
-
-export const extractAllVoiceprints = async (
-  recordingId: RecordingId,
-): Promise<BatchVoiceprintResponse> => {
-  const response = await api.post<BatchVoiceprintResponse>(
-    `/speakers/recordings/${recordingId}/voiceprints/extract-all`,
-  );
-  return response.data;
-};
-
-// Batch Operations
-export const batchArchiveRecordings = async (ids: RecordingId[]): Promise<void> => {
-  await api.post("/recordings/batch/archive", { recording_ids: ids });
-};
-
-export const batchRestoreRecordings = async (ids: RecordingId[]): Promise<void> => {
-  await api.post("/recordings/batch/restore", { recording_ids: ids });
-};
-
-export const batchSoftDeleteRecordings = async (
-  ids: RecordingId[],
-): Promise<void> => {
-  await api.post("/recordings/batch/soft-delete", { recording_ids: ids });
-};
-
-export const batchPermanentlyDeleteRecordings = async (
-  ids: RecordingId[],
-): Promise<void> => {
-  await api.post("/recordings/batch/permanent", { recording_ids: ids });
-};
-
-export const batchAddTagToRecordings = async (
-  ids: RecordingId[],
-  tagName: string,
-): Promise<void> => {
-  await api.post("/tags/batch/add", { recording_ids: ids, tag_name: tagName });
-};
-
-export const batchRemoveTagFromRecordings = async (
-  ids: RecordingId[],
-  tagName: string,
-): Promise<void> => {
-  await api.post("/tags/batch/remove", {
-    recording_ids: ids,
-    tag_name: tagName,
-  });
-};
-
-// System Setup
-export const getSystemStatus = async (): Promise<{ initialized: boolean }> => {
-  const response = await api.get<{ initialized: boolean }>("/system/status");
-  return response.data;
-};
-
-export const setupSystem = async (
-  data: {
-    username: string;
-    password: string;
-    selected_model?: string;
-  },
-  bootstrapPassword?: string,
-): Promise<{ initialized: boolean; model_preparation_task_id?: string | null }> => {
-  const response = await api.post<{ initialized: boolean; model_preparation_task_id?: string | null }>(
-    "/system/setup",
-    data,
-    buildFirstRunRequestConfig(bootstrapPassword),
-  );
-  return response.data;
-};
-
-export const checkFFmpeg = async (): Promise<{
-  ffmpeg: boolean;
-  ffprobe: boolean;
-  ffmpeg_path: string | null;
-  ffprobe_path: string | null;
-}> => {
-  const response = await api.get("/system/check-ffmpeg");
-  return response.data;
-};
-
-export const getTlsFingerprint = async (): Promise<{ fingerprint: string | null }> => {
-  try {
-    const response = await api.get<{ fingerprint: string | null }>("/system/fingerprint");
-    return response.data;
-
-    } catch (error: unknown) {
-    console.error("Failed to fetch TLS fingerprint", error);
-    return { fingerprint: null };
-  }
-};
-
-export const getRecordingInfo = async (
-  recordingId: RecordingId,
-): Promise<RecordingInfo> => {
-  const response = await api.get<RecordingInfo>(`/recordings/${recordingId}/info`);
-  return response.data;
-};
-
-export const getDemoRecording = async (): Promise<{ id: RecordingId | null }> => {
-  const response = await api.get<{ id: RecordingId | null }>(
-    "/system/demo-recording",
-  );
-  return response.data;
-};
-
-export const getTaskStatus = async (taskId: string): Promise<AsyncTaskStatus> => {
-  const response = await api.get<AsyncTaskStatus>(`/system/tasks/${taskId}`);
-  return response.data;
-};
-
-export const getInitialConfig = async (bootstrapPassword?: string): Promise<{
-  llm_provider?: string;
-  gemini_api_key?: string;
-  openai_api_key?: string;
-  anthropic_api_key?: string;
-  ollama_api_url?: string;
-  hf_token?: string;
-  selected_model?: string;
-  pyannote_models_ready?: boolean;
-  bundled_pyannote_models_ready?: boolean;
-}> => {
-  const response = await api.get(
-    "/setup/initial-config",
-    buildFirstRunRequestConfig(bootstrapPassword),
-  );
-  return response.data;
-};
-
-export const validateLLM = async (
-  provider: string,
-  apiKey: string,
-  apiUrl?: string,
-  model?: string,
-  bootstrapPassword?: string,
-): Promise<{ valid: boolean; message?: string; models?: string[] }> => {
-  const response = await api.post<{
-    valid: boolean;
-    message?: string;
-    models?: string[];
-  }>(
-    "/setup/validate-llm",
-    {
-      provider,
-      api_key: apiKey,
-      api_url: apiUrl,
-      model,
-    },
-    buildFirstRunRequestConfig(bootstrapPassword),
-  );
-  return response.data;
-};
-
-export const validateHF = async (
-  token: string,
-  bootstrapPassword?: string,
-): Promise<{ valid: boolean; message?: string }> => {
-  const response = await api.post<{ valid: boolean; message?: string }>(
-    "/setup/validate-hf",
-    { token },
-    buildFirstRunRequestConfig(bootstrapPassword),
-  );
-  return response.data;
-};
-
-export const getModelsStatus = async (
-  whisperModelSize?: string,
-): Promise<SystemModelStatus> => {
-  const params = new URLSearchParams();
-  if (whisperModelSize) params.append("whisper_model_size", whisperModelSize);
-  const response = await api.get<SystemModelStatus>(
-    `/system/models/status?${params.toString()}`,
-  );
-  return response.data;
-};
-
-export const getDownloadProgress = async (): Promise<DownloadProgress> => {
-  const response = await api.get<DownloadProgress>("/system/download-progress");
-  return response.data;
-};
-
-export const getAdminHealth = async (): Promise<AdminHealthStatus> => {
-  const response = await api.get<AdminHealthStatus>("/system/admin-health");
-  return response.data;
-};
-
-export const deleteModel = async (
-  modelId: string,
-  whisperModelSize?: string,
-): Promise<void> => {
-  const params = new URLSearchParams();
-  if (whisperModelSize) params.append("variant", whisperModelSize);
-  await api.delete(`/system/models/${modelId}?${params.toString()}`);
-};
-
-// User Management
-export const getUsers = async (
-  skip = 0,
-  limit = 100,
-  search = "",
-): Promise<{ items: User[]; total: number }> => {
-  const params = new URLSearchParams({
-    skip: skip.toString(),
-    limit: limit.toString(),
-  });
-  if (search) params.append("search", search);
-
-  const response = await api.get<{ items: User[]; total: number }>(
-    `/users?${params.toString()}`,
-  );
-  return response.data;
-};
-
-export const getUserMe = async (): Promise<User> => {
-  const response = await api.get<User>("/users/me");
-  return response.data;
-};
-
-export const updateUserMe = async (data: {
-  username?: string;
-}): Promise<User> => {
-  const response = await api.put<User>("/users/me", data);
-  return response.data;
-};
-
-export const updatePasswordMe = async (data: {
-  current_password: string;
-  new_password: string;
-}): Promise<void> => {
-  await api.put("/users/me/password", data);
-};
-
-export const createUser = async (data: {
-  username: string;
-  password: string;
-  role: string;
-}): Promise<User> => {
-  const response = await api.post<User>("/users/", data);
-  return response.data;
-};
-
-export const updateUser = async (
-  userId: number,
-  data: {
-    username?: string;
-    password?: string;
-    role?: string;
-    is_active?: boolean;
-  },
-): Promise<User> => {
-  const response = await api.patch<User>(`/users/${userId}`, data);
-  return response.data;
-};
-
-export const updateUserRole = async (
-  userId: number,
-  role: string,
-): Promise<User> => {
-  const response = await api.patch<User>(`/users/${userId}/role`, { role });
-  return response.data;
-};
-
-export const deleteUser = async (userId: number): Promise<User> => {
-  const response = await api.delete<User>(`/users/${userId}`);
-  return response.data;
-};
-
-// Invitations
-export const getInvitations = async (): Promise<Invitation[]> => {
-  const response = await api.get<Invitation[]>("/invitations/");
-  return response.data;
-};
-
-export const createInvitation = async (
-  role: string,
-  expires_in_days: number,
-  max_uses: number,
-): Promise<Invitation> => {
-  const response = await api.post<Invitation>("/invitations/", {
-    role,
-    expires_in_days,
-    max_uses,
-  });
-  return response.data;
-};
-
-export const revokeInvitation = async (id: number): Promise<Invitation> => {
-  const response = await api.post<Invitation>(`/invitations/${id}/revoke`);
-  return response.data;
-};
-
-export const deleteInvitation = async (id: number): Promise<Invitation> => {
-  const response = await api.delete<Invitation>(`/invitations/${id}`);
-  return response.data;
-};
-
-export const validateInvitation = async (
-  code: string,
-): Promise<{ valid: boolean; role: string; inviter?: string }> => {
-  const response = await api.get<{
-    valid: boolean;
-    role: string;
-    inviter?: string;
-  }>(`/invitations/validate/${code}`);
-  return response.data;
-};
-
-export const registerUser = async (
-  username: string,
-  password: string,
-  invite_code: string,
-): Promise<User> => {
-  const payload = {
-    username,
-    password,
-    invite_code,
-  };
-  const response = await api.post<User>("/users/register", payload);
-  return response.data;
-};
-
-export interface Document {
-  id: number;
-  recording_id: RecordingId;
-  title: string;
-  file_path: string;
-  file_type: string;
-  status: "PENDING" | "PROCESSING" | "READY" | "ERROR";
-  error_message?: string;
-  created_at: string;
-}
-
-export const getDocuments = async (
-  recordingId: RecordingId,
-): Promise<Document[]> => {
-  const response = await api.get<Document[]>(
-    `/recordings/${recordingId}/documents`,
-  );
-  return response.data;
-};
-
-export const uploadDocument = async (
-  recordingId: RecordingId,
-  file: File,
-): Promise<Document> => {
-  const formData = new FormData();
-  formData.append("file", file);
-  const response = await api.post<Document>(
-    `/recordings/${recordingId}/documents`,
-    formData,
-    {
-      headers: {
-        "Content-Type": "multipart/form-data",
-      },
-    },
-  );
-  return response.data;
-};
-
-export const deleteDocument = async (documentId: number): Promise<void> => {
-  await api.delete(`/documents/${documentId}`);
-};
-
-// Chat API
-export const getChatHistory = async (
-  recordingId: RecordingId,
-): Promise<ChatMessage[]> => {
-  const response = await api.get<ChatMessage[]>(
-    `/transcripts/${recordingId}/chat`,
-  );
-  return response.data;
-};
-
-export const clearChatHistory = async (recordingId: RecordingId): Promise<void> => {
-  await api.delete(`/transcripts/${recordingId}/chat`);
-};
-
-export const streamChatMessage = (
-  recordingId: RecordingId,
-  message: string,
-  onToken: (token: string) => void,
-  onComplete: () => void,
-  onError: (error: string) => void,
-  tagIds?: number[],
-  onNotesUpdate?: () => void,
-): AbortController => {
-  const controller = new AbortController();
-
-  fetch(`${API_BASE_URL}/transcripts/${recordingId}/chat`, {
-    method: "POST",
-    credentials: "include",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify({ message, tag_ids: tagIds }),
-    signal: controller.signal,
-  })
-    .then(async (response) => {
-      if (!response.ok) {
-        try {
-          const err = await response.json();
-          onError(err.detail || "Failed to send message");
-        } catch {
-          onError(`Error: ${response.statusText}`);
-        }
-        return;
-      }
-
-      if (!response.body) {
-        onError("No response body");
-        return;
-      }
-
-      const reader = response.body.getReader();
-      const decoder = new TextDecoder();
-      let pendingLine = "";
-
-      while (true) {
-        const { done, value } = await reader.read();
-        if (done) {
-          if (pendingLine.trim()) {
-            handleChatStreamLine(pendingLine, onToken, onError, onNotesUpdate);
-          }
-          onComplete();
-          break;
-        }
-
-        const chunk = decoder.decode(value, { stream: true });
-        const lines = `${pendingLine}${chunk}`.split("\n");
-        pendingLine = lines.pop() ?? "";
-
-        for (const line of lines) {
-          handleChatStreamLine(line, onToken, onError, onNotesUpdate);
-        }
-      }
-    })
-    .catch((err) => {
-      if (isAbortError(err)) {
-        return;
-      } else {
-        onError(getErrorMessage(err, "Network error"));
-      }
-    });
-
-  return controller;
-};
-
-const handleChatStreamLine = (
-  line: string,
-  onToken: (token: string) => void,
-  onError: (error: string) => void,
-  onNotesUpdate?: () => void,
-) => {
-  if (line.startsWith("event: notes_update")) {
-    onNotesUpdate?.();
-    return;
-  }
-
-  if (!line.startsWith("data: ")) {
-    return;
-  }
-
-  const data = line.slice(6);
-  if (data === "[DONE]") {
-    return;
-  }
-
-  try {
-    const parsed = JSON.parse(data);
-    if (parsed.token !== undefined) {
-      onToken(parsed.token);
-    } else if (parsed.error) {
-      onError(parsed.error);
-    }
-  } catch (error) {
-    console.error("Failed to parse SSE data", error);
-  }
-};
-
-// Infer Speakers
-export const inferSpeakers = async (recordingId: RecordingId): Promise<void> => {
-  await api.post(`/recordings/${recordingId}/infer-speakers`);
-};
-
-export const listModels = async (
-  provider: string,
-  apiKey: string,
-  apiUrl?: string,
-  bootstrapPassword?: string,
-): Promise<{ models: string[] }> => {
-  const response = await api.post<{ models: string[] }>(
-    "/setup/list-models",
-    {
-      provider,
-      api_key: apiKey,
-      api_url: apiUrl,
-    },
-    buildFirstRunRequestConfig(bootstrapPassword),
-  );
-  return response.data;
-};
-
-export const fetchProxyModels = async (
-  provider: string,
-  apiUrl?: string,
-  apiKey?: string,
-): Promise<string[]> => {
-  const params = new URLSearchParams();
-  params.append("provider", provider);
-  if (apiUrl) params.append("api_url", apiUrl);
-  if (apiKey) params.append("api_key", apiKey);
-
-  const response = await api.get<string[]>(`/llm/models?${params.toString()}`);
-  return response.data;
-};
-
-// Backup & Restore
-export const exportBackupAsync = async (
-  includeAudio: boolean = true,
-): Promise<{ task_id: string }> => {
-  const response = await api.post<{ task_id: string }>(
-    `/backup/export?include_audio=${includeAudio}`,
-  );
-  return response.data;
-};
-
-export const getBackupStatus = async (
-  taskId: string,
-): Promise<{ state: string; status: string; result?: unknown }> => {
-  const response = await api.get<{
-    state: string;
-    status: string;
-    result?: unknown;
-  }>(`/backup/export/${taskId}`);
-  return response.data;
-};
-
-export const downloadBackupFile = async (taskId: string): Promise<void> => {
-  const response = await api.get(`/backup/export/${taskId}/download`, {
-    responseType: "blob",
-  });
-
-  const blob = new Blob([response.data], { type: "application/zip" });
-  const url = window.URL.createObjectURL(blob);
-  const a = document.createElement("a");
-  a.href = url;
-
-  // Create filename with timestamp
-  const filename = `nojoin_backup_${new Date().toISOString().replace(/[:.]/g, "-")}.zip`;
-  a.download = filename;
-
-  document.body.appendChild(a);
-  a.click();
-  window.URL.revokeObjectURL(url);
-  document.body.removeChild(a);
-};
-
-export const importBackup = async (
-  file: File,
-  clearExisting: boolean,
-  overwriteExisting: boolean,
-  onProgress?: (progress: number) => void,
-  onStatus?: (status: string) => void,
-): Promise<void> => {
-  const formData = new FormData();
-  formData.append("file", file);
-
-  try {
-    // 1. Upload and Start Job
-    const response = await api.post<{ job_id?: string }>(
-      `/backup/import?clear_existing=${clearExisting}&overwrite_existing=${overwriteExisting}`,
-      formData,
-      {
-        headers: {
-          "Content-Type": "multipart/form-data",
-        },
-        timeout: 300000, // 5 minutes for upload (large files)
-        onUploadProgress: (progressEvent) => {
-          if (onProgress && progressEvent.total) {
-            const percentCompleted = Math.round(
-              (progressEvent.loaded * 100) / progressEvent.total,
-            );
-            // Cap at 99% during upload, save 100% for completion or processing start
-            onProgress(Math.min(percentCompleted, 99));
-          }
-        },
-      },
-    );
-
-    // A 202 response carries a job_id to poll; a response without one means the
-    // import already completed synchronously.
-    const jobId = response.data.job_id;
-
-    if (jobId) {
-      if (onProgress) onProgress(100);
-      if (onStatus) onStatus("Processing on server...");
-
-      // Poll for status
-      return new Promise<void>((resolve, reject) => {
-        const pollInterval = setInterval(async () => {
-          try {
-            const statusRes = await api.get(`/backup/import/${jobId}`);
-            const { status, progress, error } = statusRes.data;
-
-            if (onStatus && progress) {
-              onStatus(progress);
-            }
-
-            if (status === "completed") {
-              clearInterval(pollInterval);
-              resolve();
-            } else if (status === "failed") {
-              clearInterval(pollInterval);
-              reject(new Error(error || "Restore failed during processing"));
-            }
-            // If 'pending' or 'processing', continue polling
-
-                    } catch (err: unknown) {
-            // Retries on transient network errors; aborts on 404 (lost job).
-            if (getErrorStatus(err) === 404) {
-              clearInterval(pollInterval);
-              reject(new Error("Restore job lost on server"));
-            }
-          }
-        }, 2000); // Poll every 2 seconds
-      });
-    }
-
-    // No job_id: the import completed synchronously, so report done.
-    if (onProgress) onProgress(100);
-    return;
-
-    } catch (error: unknown) {
-    throw error;
-  }
-};
-
-interface UploadChunkInitResponse {
-  upload_id: string;
-}
-
-export const uploadBackupChunked = async (
-  file: File,
-  clearExisting: boolean,
-  overwriteExisting: boolean,
-  onProgress: (percent: number) => void,
-  onStatus: (status: string) => void,
-): Promise<void> => {
-  // 10MB chunks to stay well under Cloudflare 100MB limit
-  const CHUNK_SIZE = 10 * 1024 * 1024;
-  const totalChunks = Math.ceil(file.size / CHUNK_SIZE);
-
-  try {
-    // 1. Initialize
-    onStatus("Initializing upload...");
-    const initRes = await api.post<UploadChunkInitResponse>(
-      "/backup/upload/init",
-      null,
-      {
-        params: {
-          filename: file.name,
-          file_size: file.size,
-          total_chunks: totalChunks,
-        },
-      },
-    );
-    const { upload_id } = initRes.data;
-
-    // 2. Upload Chunks
-    for (let i = 0; i < totalChunks; i++) {
-      const start = i * CHUNK_SIZE;
-      const end = Math.min(start + CHUNK_SIZE, file.size);
-      const chunk = file.slice(start, end);
-
-      const formData = new FormData();
-      // blob name is important
-      formData.append("file", chunk, "blob");
-
-      await api.post(`/backup/upload/${upload_id}/chunk`, formData, {
-        params: { chunk_index: i },
-        headers: { "Content-Type": "multipart/form-data" },
-      });
-
-      // Update combined progress
-      const percent = Math.round(((i + 1) / totalChunks) * 100);
-      onProgress(Math.min(percent, 99)); // Keep 100 for processing
-      onStatus(`Uploading part ${i + 1} of ${totalChunks}...`);
-    }
-
-    // 3. Complete
-    onStatus("Finalizing upload and starting restore...");
-    const completeRes = await api.post(
-      `/backup/upload/${upload_id}/complete`,
-      null,
-      {
-        params: {
-          clear_existing: clearExisting,
-          overwrite_existing: overwriteExisting,
-        },
-      },
-    );
-
-    // 4. Poll Status (Reusing logic pattern from importBackup)
-    const jobId = completeRes.data.job_id;
-    if (jobId) {
-      if (onProgress) onProgress(100);
-      if (onStatus) onStatus("Processing on server...");
-
-      return new Promise<void>((resolve, reject) => {
-        const pollInterval = setInterval(async () => {
-          try {
-            const statusRes = await api.get(`/backup/import/${jobId}`);
-            const { status, progress, error } = statusRes.data;
-
-            if (onStatus && progress) {
-              onStatus(progress);
-            }
-
-            if (status === "completed") {
-              clearInterval(pollInterval);
-              resolve();
-            } else if (status === "failed") {
-              clearInterval(pollInterval);
-              reject(new Error(error || "Restore failed during processing"));
-            }
-
-                    } catch (err: unknown) {
-            console.warn("Polling status failed", err);
-            // If 404, job lost?
-            if (getErrorStatus(err) === 404) {
-              clearInterval(pollInterval);
-              reject(new Error("Restore job lost on server"));
-            }
-          }
-        }, 2000);
-      });
-    }
-
-    } catch (error: unknown) {
-    throw error;
-  }
-};
-
-export const seedDemoData = async (): Promise<void> => {
-  await api.post("/system/seed-demo");
-};
-
-export const getVersion = async (
-  options?: { refresh?: boolean },
-): Promise<VersionInfo> => {
-  const response = await api.get<VersionInfo>("/version", {
-    params: options?.refresh ? { refresh: true } : undefined,
-  });
-  return response.data;
-};
-
-export const getSpeakerSegments = async (
-  speakerId: number,
-): Promise<SpeakerSegment[]> => {
-  const response = await api.get<SpeakerSegment[]>(
-    `/speakers/${speakerId}/segments`,
-  );
-  return response.data;
-};
-
-export const recalibrateSpeaker = async (
-  speakerId: number,
-  segments: SegmentSelection[],
-): Promise<void> => {
-  await api.post(`/speakers/${speakerId}/recalibrate`, segments);
-};
-
-export const splitSpeaker = async (
-  speakerId: number,
-  newSpeakerName: string,
-  segments: SegmentSelection[],
-): Promise<GlobalSpeaker> => {
-  const response = await api.post<GlobalSpeaker>(
-    `/speakers/${speakerId}/split`,
-    {
-      new_speaker_name: newSpeakerName,
-      segments,
-    },
-  );
-  return response.data;
-};
-
-export const splitLocalSpeaker = async (
-  recordingId: RecordingId,
-  diarizationLabel: string,
-  newSpeakerName: string,
-  segments: SegmentSelection[],
-) => {
-  const response = await api.post(
-    `/speakers/recordings/${recordingId}/speakers/${diarizationLabel}/split`,
-    {
-      new_speaker_name: newSpeakerName,
-      segments,
-    },
-  );
-  return response.data;
-};
-
-export const scanMatches = async (
-  speakerId: number,
-): Promise<{ matches_found: number; recordings_updated: number }> => {
-  const response = await api.post(`/speakers/${speakerId}/scan-matches`);
-  return response.data;
-};
-
-export default api;
